+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - Improved console command messaging when a component is disabled.
+- Changed the token request parameter check to use the value defined in the general config settings ([#307](https://github.com/putyourlightson/craft-blitz/issues/307)).
 
 ### Fixed
 - Fixed an error that could occur when an empty response was encountered with the local warmer ([#305](https://github.com/putyourlightson/craft-blitz/issues/305)).